Add support to launch instance into a VPC subnet
Add a rudimentary support for launching instances into VPC. Mainly
this means new vpc_subnet_id argument which specifies the
"subnet-xxx" id of the Amazon EC2 VPC subnet. As the subnet can
live only in one Availability Zone then the availability zone argument
is not needed when launching into vpc. Because of this there's
a new way to specify an array of arguments where only one argument
is required.

This patch also refactored away the Launcher::verify_required_options
method as the required options are already checked on the
metadata_launcher.

This commit contains validation fixes suggested in this github discussion
https://github.com/airbnb/stemcell/pull/46