Add support to launch instance into a VPC subnet
Add a rudimentary support for launching instances into VPC. Mainly
this means new vpc_subnet_id argument which specifies the
"subnet-xxx" id of the Amazon EC2 VPC subnet. As the subnet can
live only in one Availability Zone then the availability zone argument
is not needed when launching into vpc. Because of this there's
a new way to specify an array of arguments where only one argument
is required. In addition there's "private_ip_address" option
which can be used to set instance private ip on launch.