Put Launcher.verify_required_options() back
This function is required to be present when stemcell
is used as an API and the Launcher is one of the main interfaces
on this api.

MetadataLauncher then uses this function to check that
all required parameters are present. Also a new function
create_launcher() is added as the Launcher class needs
to be instantiased before validate_options can call the
verify_required_options() function.