improve batch operation in launcher
Summary:

We relied stemcell heavily in our web UI. However, we observed that
some of the operation returns very obsecure error message.  For
example, when we run `launch!` and the instances failed to run
`set_tags`, caller will receive an exception saying that one of the
instance failed, with no details of the other successfully launched
instances.

In this diff, I systematically improved the "batch operation" against
the instances, which
  1. run operation on a batch of instances with more flexibly
  2. provide informative error message about instance statuses when
     error occurs.

Test Plan:

Added new spec to make sure `run_batch_operation` worked as expected.