[launcher] wait for running state to set classic link
we cannot enable classic link unless the instance is in a running state. we
assume that if classic link is desired, then we should take the time to wait
even if the wait flag was not specified. we apply classic link as instances
become running, instead of waiting for them all to become running, as an
optimization to reduce wait times.

we also refactor the wait time to reduce state, and make the magic number in
the code a constant.